Fix Text-to-Speech
The existing TTS method was broken because the `audio` DOM-element which
was being rendered relied on Google serving back an audio file when a
query is made against the Google TTS endpoint. This endpoint now has a
captcha by default.

Instead `SpeechSynthesisUtterance` is used to request TTS from the
Chrome browser without the need for a direct URL request.